move core.py to core/__init__.py
